fix it was not checked if it was actually locked, fix ID was hashed even when not needed, simplify code by reusing existing logic
